test: structure ramps mocks
