+++ tests/api-mocking/mock-responses/defaults/onramp-apis.ts
@@ -3,13 +3,11 @@ import {
   RampsRegions,
   RampsRegionsEnum,
 } from '../../../framework/Constants.ts';
-import {
-  RAMPS_NETWORKS_RESPONSE,
-  RAMPS_TOP_TOKENS_RESPONSE,
-  RAMPS_PROVIDERS_RESPONSE,
-  RAMPS_PAYMENTS_V2_RESPONSE,
-} from '../ramps/ramps-mocks.ts';
-import { createGeolocationResponse } from '../ramps/ramps-geolocation.ts';
+import { RAMPS_NETWORKS_RESPONSE } from '../ramps/responses/ramps-networks-response.ts';
+import { RAMPS_TOP_TOKENS_RESPONSE } from '../ramps/responses/ramps-tokens-response.ts';
+import { RAMPS_PROVIDERS_RESPONSE } from '../ramps/responses/ramps-providers-response.ts';
+import { RAMPS_PAYMENTS_V2_RESPONSE } from '../ramps/responses/ramps-payments-response.ts';
+import { createGeolocationResponse } from '../ramps/responses/ramps-geolocation.ts';
 
 /**
  * Mock data for on-ramp API endpoints used in E2E testing.
