--- tests/api-mocking/mock-responses/ramps/ramps-quotes-response.ts
@@ -1,2238 +1,0 @@
-/* eslint-disable @metamask/design-tokens/color-no-hex */
-export const RAMPS_QUOTE_RESPONSE = {
-  success: [
-    {
-      provider: '/providers/mercuryo',
-      url: null,
-      method: null,
-      headers: null,
-      quote: {
-        crypto: {
-          id: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          idv2: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          legacyId: '/currencies/crypto/1/eth',
-          network: {
-            active: true,
-            chainId: '1',
-            chainName: 'Ethereum Mainnet',
-            shortName: 'Ethereum',
-          },
-          logo: 'https://uat-static.cx.metamask.io/api/v1/tokenIcons/1/0x0000000000000000000000000000000000000000.png',
-          decimals: 18,
-          address: '0x0000000000000000000000000000000000000000',
-          symbol: 'ETH',
-          name: 'Ethereum',
-        },
-        cryptoId:
-          '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-        fiat: {
-          id: '/currencies/fiat/eur',
-          symbol: 'EUR',
-          name: 'Euro',
-          decimals: 2,
-          denomSymbol: '€',
-        },
-        fiatId: '/currencies/fiat/eur',
-        amountIn: 100,
-        amountOut: 0.024581104642021276,
-        exchangeRate: 4014.8724574114517,
-        networkFee: 0.03,
-        providerFee: 1.28,
-        extraFee: 0,
-        receiver: '0x76cf1CdD1fcC252442b50D6e97207228aA4aefC3',
-        paymentMethod: '/payments/apple-pay',
-        cryptoTranslation: 'ETH_1',
-      },
-      nativeApplePay: {},
-      providerInfo: {
-        id: '/providers/mercuryo',
-        name: 'Mercuryo',
-        environmentType: 'PRODUCTION',
-        description:
-          'Per Mercuryo: “Mercuryo offers easy onboarding for MetaMask users, with a speedy purchase process of under 15 seconds Light KYC up to 700$. With support for 20+ tokens, customers can pay using preferred methods, such as Apple Pay and bank cards.”',
-        hqAddress: 'London, United Kingdom, 77 Gracechurch, EC3V0AG',
-        links: [
-          {
-            name: 'Homepage',
-            url: 'https://mercuryo.io/',
-          },
-          {
-            name: 'Privacy Policy',
-            url: 'https://mercuryo.io/legal/privacy/',
-          },
-          {
-            name: 'Support',
-            url: 'https://help.mercuryo.io/en/',
-          },
-        ],
-        logos: {
-          light: '/assets/providers/mercuryo_light.png',
-          dark: '/assets/providers/mercuryo_dark.png',
-          height: 24,
-          width: 88,
-        },
-        features: {
-          buy: {
-            enabled: true,
-            userAgent: null,
-            padCustomOrderId: false,
-            orderCustomId: '',
-            browser: 'APP_BROWSER',
-            orderCustomIdRequired: false,
-            orderCustomIdExpiration: null,
-            orderCustomIdSeparator: null,
-            orderCustomIdPrefixes: ['c-', ''],
-            supportedByBackend: true,
-            redirection: 'JSON_REDIRECTION',
-          },
-          quotes: {
-            enabled: false,
-            supportedByBackend: false,
-          },
-          sell: {
-            enabled: true,
-          },
-          sellQuotes: {
-            enabled: true,
-          },
-          recurringBuy: {
-            enabled: true,
-          },
-        },
-      },
-      providerQuote: {
-        paymentMethod: ['apple-pay'],
-        fiatCurrency: 'EUR',
-        cryptoCurrency: 'ETH_1',
-      },
-      metadata: {
-        reliability: 0.807440599,
-        tags: {
-          isMostReliable: true,
-        },
-      },
-    },
-    {
-      provider: '/providers/mercuryo-staging',
-      url: null,
-      method: null,
-      headers: null,
-      quote: {
-        crypto: {
-          id: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          idv2: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          legacyId: '/currencies/crypto/1/eth',
-          network: {
-            active: true,
-            chainId: '1',
-            chainName: 'Ethereum Mainnet',
-            shortName: 'Ethereum',
-          },
-          logo: 'https://uat-static.cx.metamask.io/api/v1/tokenIcons/1/0x0000000000000000000000000000000000000000.png',
-          decimals: 18,
-          address: '0x0000000000000000000000000000000000000000',
-          symbol: 'ETH',
-          name: 'Ethereum',
-        },
-        cryptoId:
-          '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-        fiat: {
-          id: '/currencies/fiat/eur',
-          symbol: 'EUR',
-          name: 'Euro',
-          decimals: 2,
-          denomSymbol: '€',
-        },
-        fiatId: '/currencies/fiat/eur',
-        amountIn: 100,
-        amountOut: 0.0231989044130259,
-        exchangeRate: 4008.8100000007603,
-        networkFee: 0.01,
-        providerFee: 6.99,
-        extraFee: 0,
-        receiver: '0x76cf1CdD1fcC252442b50D6e97207228aA4aefC3',
-        paymentMethod: '/payments/apple-pay',
-        cryptoTranslation: 'ETH_1',
-      },
-      nativeApplePay: {},
-      providerInfo: {
-        id: '/providers/mercuryo-staging',
-        name: 'Mercuryo (Staging)',
-        environmentType: 'STAGING',
-        description:
-          'Per Mercuryo: “Mercuryo offers easy onboarding for MetaMask users, with a speedy purchase process of under 15 seconds Light KYC up to 700$. With support for 20+ tokens, customers can pay using preferred methods, such as Apple Pay and bank cards.”',
-        hqAddress: 'London, United Kingdom, 77 Gracechurch, EC3V0AG',
-        links: [
-          {
-            name: 'Homepage',
-            url: 'https://mercuryo.io/',
-          },
-          {
-            name: 'Privacy Policy',
-            url: 'https://mercuryo.io/legal/privacy/',
-          },
-          {
-            name: 'Support',
-            url: 'https://help.mercuryo.io/en/',
-          },
-        ],
-        logos: {
-          light: '/assets/providers/mercuryo_light.png',
-          dark: '/assets/providers/mercuryo_dark.png',
-          height: 24,
-          width: 88,
-        },
-        features: {
-          buy: {
-            enabled: true,
-            userAgent: null,
-            padCustomOrderId: false,
-            orderCustomId: '',
-            browser: 'APP_BROWSER',
-            orderCustomIdRequired: false,
-            orderCustomIdExpiration: null,
-            orderCustomIdSeparator: null,
-            orderCustomIdPrefixes: ['c-', ''],
-            supportedByBackend: true,
-            redirection: 'JSON_REDIRECTION',
-          },
-          quotes: {
-            enabled: false,
-            supportedByBackend: false,
-          },
-          sell: {
-            enabled: true,
-          },
-          sellQuotes: {
-            enabled: true,
-          },
-          recurringBuy: {
-            enabled: true,
-          },
-        },
-      },
-      providerQuote: {
-        paymentMethod: ['apple-pay'],
-        fiatCurrency: 'EUR',
-        cryptoCurrency: 'ETH_1',
-      },
-      metadata: {
-        reliability: 0.807440599,
-        tags: {},
-      },
-    },
-    {
-      provider: '/providers/moonpay-staging',
-      url: null,
-      method: null,
-      headers: null,
-      quote: {
-        crypto: {
-          id: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          idv2: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          legacyId: '/currencies/crypto/1/eth',
-          network: {
-            active: true,
-            chainId: '1',
-            chainName: 'Ethereum Mainnet',
-            shortName: 'Ethereum',
-          },
-          logo: 'https://uat-static.cx.metamask.io/api/v1/tokenIcons/1/0x0000000000000000000000000000000000000000.png',
-          decimals: 18,
-          address: '0x0000000000000000000000000000000000000000',
-          symbol: 'ETH',
-          name: 'Ethereum',
-        },
-        cryptoId:
-          '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-        fiat: {
-          id: '/currencies/fiat/eur',
-          symbol: 'EUR',
-          name: 'Euro',
-          decimals: 2,
-          denomSymbol: '€',
-        },
-        fiatId: '/currencies/fiat/eur',
-        amountIn: 100,
-        amountOut: 0.0248,
-        exchangeRate: 3851.209677419355,
-        networkFee: 0.9,
-        providerFee: 3.59,
-        extraFee: 0,
-        receiver: '0x76cf1CdD1fcC252442b50D6e97207228aA4aefC3',
-        paymentMethod: '/payments/apple-pay',
-        cryptoTranslation: 'eth',
-      },
-      nativeApplePay: {
-        supported: false,
-        merchantId: '',
-      },
-      providerInfo: {
-        id: '/providers/moonpay-staging',
-        name: 'Moonpay (Staging)',
-        environmentType: 'STAGING',
-        description:
-          'Per MoonPay: “MoonPay provides a smooth experience for converting between fiat currencies and cryptocurrencies. Easily top-up with ETH, BNB and more directly in your MetaMask wallet via MoonPay using all major payment methods including debit and credit card, local bank transfers, Apple Pay, Google Pay, and Samsung Pay. MoonPay is active in more than 160 countries and is trusted by 250+ leading wallets, websites, and applications.”',
-        hqAddress: '8 The Green, Dover, DE, 19901, USA',
-        links: [
-          {
-            name: 'Homepage',
-            url: 'https://www.moonpay.com/',
-          },
-          {
-            name: 'Privacy Policy',
-            url: 'https://www.moonpay.com/legal/privacy_policy',
-          },
-          {
-            name: 'Support',
-            url: 'https://support.moonpay.com/hc/en-gb/categories/360001595097-Customer-Support-Help-Center',
-          },
-        ],
-        logos: {
-          light: '/assets/providers/moonpay_light.png',
-          dark: '/assets/providers/moonpay_dark.png',
-          height: 24,
-          width: 88,
-        },
-        features: {
-          buy: {
-            enabled: true,
-            userAgent: null,
-            padCustomOrderId: false,
-            orderCustomId: 'GUID',
-            browser: 'APP_BROWSER',
-            orderCustomIdRequired: true,
-            orderCustomIdExpiration: 60,
-            orderCustomIdSeparator: null,
-            orderCustomIdPrefixes: ['c-', ''],
-            supportedByBackend: true,
-            redirection: 'JSON_REDIRECTION',
-          },
-          quotes: {
-            enabled: false,
-            supportedByBackend: false,
-          },
-          sell: {
-            enabled: true,
-          },
-          sellQuotes: {
-            enabled: true,
-          },
-          recurringBuy: {
-            enabled: false,
-          },
-        },
-      },
-      providerQuote: {
-        paymentMethod: ['apple-pay'],
-        fiatCurrency: 'EUR',
-        cryptoCurrency: 'eth',
-      },
-      metadata: {
-        reliability: 0.6823529412,
-        tags: {},
-      },
-    },
-    {
-      provider: '/providers/moonpay',
-      url: null,
-      method: null,
-      headers: null,
-      quote: {
-        crypto: {
-          id: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          idv2: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          legacyId: '/currencies/crypto/1/eth',
-          network: {
-            active: true,
-            chainId: '1',
-            chainName: 'Ethereum Mainnet',
-            shortName: 'Ethereum',
-          },
-          logo: 'https://uat-static.cx.metamask.io/api/v1/tokenIcons/1/0x0000000000000000000000000000000000000000.png',
-          decimals: 18,
-          address: '0x0000000000000000000000000000000000000000',
-          symbol: 'ETH',
-          name: 'Ethereum',
-        },
-        cryptoId:
-          '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-        fiat: {
-          id: '/currencies/fiat/eur',
-          symbol: 'EUR',
-          name: 'Euro',
-          decimals: 2,
-          denomSymbol: '€',
-        },
-        fiatId: '/currencies/fiat/eur',
-        amountIn: 100,
-        amountOut: 0.0248,
-        exchangeRate: 3851.209677419355,
-        networkFee: 0.9,
-        providerFee: 3.59,
-        extraFee: 0,
-        receiver: '0x76cf1CdD1fcC252442b50D6e97207228aA4aefC3',
-        paymentMethod: '/payments/apple-pay',
-        cryptoTranslation: 'eth',
-      },
-      nativeApplePay: {
-        supported: false,
-        merchantId: '',
-      },
-      providerInfo: {
-        id: '/providers/moonpay',
-        name: 'Moonpay',
-        environmentType: 'PRODUCTION',
-        description:
-          'Per MoonPay: “MoonPay provides a smooth experience for converting between fiat currencies and cryptocurrencies. Easily top-up with ETH, BNB and more directly in your MetaMask wallet via MoonPay using all major payment methods including debit and credit card, local bank transfers, Apple Pay, Google Pay, and Samsung Pay. MoonPay is active in more than 160 countries and is trusted by 250+ leading wallets, websites, and applications.”',
-        hqAddress: '8 The Green, Dover, DE, 19901, USA',
-        links: [
-          {
-            name: 'Homepage',
-            url: 'https://www.moonpay.com/',
-          },
-          {
-            name: 'Privacy Policy',
-            url: 'https://www.moonpay.com/legal/privacy_policy',
-          },
-          {
-            name: 'Support',
-            url: 'https://support.moonpay.com/hc/en-gb/categories/360001595097-Customer-Support-Help-Center',
-          },
-        ],
-        logos: {
-          light: '/assets/providers/moonpay_light.png',
-          dark: '/assets/providers/moonpay_dark.png',
-          height: 24,
-          width: 88,
-        },
-        features: {
-          buy: {
-            enabled: true,
-            userAgent: null,
-            padCustomOrderId: false,
-            orderCustomId: 'GUID',
-            browser: 'APP_BROWSER',
-            orderCustomIdRequired: true,
-            orderCustomIdExpiration: 60,
-            orderCustomIdSeparator: null,
-            orderCustomIdPrefixes: ['c-', ''],
-            supportedByBackend: true,
-            redirection: 'JSON_REDIRECTION',
-          },
-          quotes: {
-            enabled: false,
-            supportedByBackend: false,
-          },
-          sell: {
-            enabled: true,
-          },
-          sellQuotes: {
-            enabled: true,
-          },
-          recurringBuy: {
-            enabled: false,
-          },
-        },
-      },
-      providerQuote: {
-        paymentMethod: ['apple-pay'],
-        fiatCurrency: 'EUR',
-        cryptoCurrency: 'eth',
-      },
-      metadata: {
-        reliability: 0.6823529412,
-        tags: {},
-      },
-    },
-    {
-      provider: '/providers/transak',
-      url: null,
-      method: null,
-      headers: null,
-      quote: {
-        crypto: {
-          id: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          idv2: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          legacyId: '/currencies/crypto/1/eth',
-          network: {
-            active: true,
-            chainId: '1',
-            chainName: 'Ethereum Mainnet',
-            shortName: 'Ethereum',
-          },
-          logo: 'https://uat-static.cx.metamask.io/api/v1/tokenIcons/1/0x0000000000000000000000000000000000000000.png',
-          decimals: 18,
-          address: '0x0000000000000000000000000000000000000000',
-          symbol: 'ETH',
-          name: 'Ethereum',
-        },
-        cryptoId:
-          '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-        fiat: {
-          id: '/currencies/fiat/eur',
-          symbol: 'EUR',
-          name: 'Euro',
-          decimals: 2,
-          denomSymbol: '€',
-        },
-        fiatId: '/currencies/fiat/eur',
-        amountIn: 100,
-        amountOut: 0.02381807,
-        exchangeRate: 4072.1183538380733,
-        networkFee: 0.02,
-        providerFee: 2.99,
-        extraFee: 0,
-        receiver: '0x76cf1CdD1fcC252442b50D6e97207228aA4aefC3',
-        paymentMethod: '/payments/apple-pay',
-        cryptoTranslation: 'ETHethereum',
-      },
-      nativeApplePay: {},
-      providerInfo: {
-        id: '/providers/transak',
-        name: 'Transak',
-        environmentType: 'PRODUCTION',
-        description:
-          'Per Transak: "The fastest and securest way to buy 100+ cryptocurrencies on 75+ blockchains. Pay via Apple Pay, UPI, bank transfer or use your debit or credit card. Trusted by 2+ million global users. Transak empowers wallets, gaming, DeFi, NFTs, Exchanges, and DAOs across 125+ countries."',
-        hqAddress:
-          '35 Shearing Street, Bury St. Edmunds, IP32 6FE, United Kingdom',
-        links: [
-          {
-            name: 'Homepage',
-            url: 'https://www.transak.com/',
-          },
-          {
-            name: 'Privacy Policy',
-            url: 'https://transak.com/privacy-policy',
-          },
-          {
-            name: 'Support',
-            url: 'https://support.transak.com/hc/en-us',
-          },
-        ],
-        logos: {
-          light: '/assets/providers/transak_light.png',
-          dark: '/assets/providers/transak_dark.png',
-          height: 24,
-          width: 90,
-        },
-        features: {
-          buy: {
-            enabled: true,
-            userAgent: null,
-            padCustomOrderId: false,
-            orderCustomId: '',
-            browser: null,
-            orderCustomIdRequired: false,
-            orderCustomIdExpiration: null,
-            orderCustomIdSeparator: null,
-            orderCustomIdPrefixes: ['c-', ''],
-            supportedByBackend: true,
-            redirection: 'JSON_REDIRECTION',
-          },
-          quotes: {
-            enabled: false,
-            supportedByBackend: false,
-          },
-          sell: {
-            enabled: true,
-          },
-          sellQuotes: {
-            enabled: true,
-          },
-          recurringBuy: {},
-        },
-      },
-      providerQuote: {
-        paymentMethod: ['apple-pay'],
-        fiatCurrency: 'EUR',
-        cryptoCurrency: 'ETHethereum',
-      },
-      metadata: {
-        reliability: 0.6578456951,
-        tags: {},
-      },
-    },
-    {
-      provider: '/providers/transak-staging',
-      url: null,
-      method: null,
-      headers: null,
-      quote: {
-        crypto: {
-          id: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          idv2: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          legacyId: '/currencies/crypto/1/eth',
-          network: {
-            active: true,
-            chainId: '1',
-            chainName: 'Ethereum Mainnet',
-            shortName: 'Ethereum',
-          },
-          logo: 'https://uat-static.cx.metamask.io/api/v1/tokenIcons/1/0x0000000000000000000000000000000000000000.png',
-          decimals: 18,
-          address: '0x0000000000000000000000000000000000000000',
-          symbol: 'ETH',
-          name: 'Ethereum',
-        },
-        cryptoId:
-          '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-        fiat: {
-          id: '/currencies/fiat/eur',
-          symbol: 'EUR',
-          name: 'Euro',
-          decimals: 2,
-          denomSymbol: '€',
-        },
-        fiatId: '/currencies/fiat/eur',
-        amountIn: 100,
-        amountOut: 0.02369152,
-        exchangeRate: 4072.34318439678,
-        networkFee: 0.02,
-        providerFee: 3.5,
-        extraFee: 0,
-        receiver: '0x76cf1CdD1fcC252442b50D6e97207228aA4aefC3',
-        paymentMethod: '/payments/apple-pay',
-        cryptoTranslation: 'ETHethereum',
-      },
-      nativeApplePay: {},
-      providerInfo: {
-        id: '/providers/transak-staging',
-        name: 'Transak (Staging)',
-        environmentType: 'STAGING',
-        description:
-          'Per Transak: "The fastest and securest way to buy 100+ cryptocurrencies on 75+ blockchains. Pay via Apple Pay, UPI, bank transfer or use your debit or credit card. Trusted by 2+ million global users. Transak empowers wallets, gaming, DeFi, NFTs, Exchanges, and DAOs across 125+ countries."',
-        hqAddress:
-          '35 Shearing Street, Bury St. Edmunds, IP32 6FE, United Kingdom',
-        links: [
-          {
-            name: 'Homepage',
-            url: 'https://www.transak.com/',
-          },
-          {
-            name: 'Privacy Policy',
-            url: 'https://transak.com/privacy-policy',
-          },
-          {
-            name: 'Support',
-            url: 'https://support.transak.com/hc/en-us',
-          },
-        ],
-        logos: {
-          light: '/assets/providers/transak_light.png',
-          dark: '/assets/providers/transak_dark.png',
-          height: 24,
-          width: 90,
-        },
-        features: {
-          buy: {
-            enabled: true,
-            userAgent: null,
-            padCustomOrderId: false,
-            orderCustomId: '',
-            browser: null,
-            orderCustomIdRequired: false,
-            orderCustomIdExpiration: null,
-            orderCustomIdSeparator: null,
-            orderCustomIdPrefixes: ['c-', ''],
-            supportedByBackend: true,
-            redirection: 'JSON_REDIRECTION',
-          },
-          quotes: {
-            enabled: false,
-            supportedByBackend: false,
-          },
-          sell: {
-            enabled: true,
-          },
-          sellQuotes: {
-            enabled: true,
-          },
-          recurringBuy: {},
-        },
-      },
-      providerQuote: {
-        paymentMethod: ['apple-pay'],
-        fiatCurrency: 'EUR',
-        cryptoCurrency: 'ETHethereum',
-      },
-      metadata: {
-        reliability: 0.6578456951,
-        tags: {},
-      },
-    },
-    {
-      provider: '/providers/banxa-staging',
-      url: null,
-      method: null,
-      headers: null,
-      quote: {
-        crypto: {
-          id: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          idv2: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          legacyId: '/currencies/crypto/1/eth',
-          network: {
-            active: true,
-            chainId: '1',
-            chainName: 'Ethereum Mainnet',
-            shortName: 'Ethereum',
-          },
-          logo: 'https://uat-static.cx.metamask.io/api/v1/tokenIcons/1/0x0000000000000000000000000000000000000000.png',
-          decimals: 18,
-          address: '0x0000000000000000000000000000000000000000',
-          symbol: 'ETH',
-          name: 'Ethereum',
-        },
-        cryptoId:
-          '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-        fiat: {
-          id: '/currencies/fiat/eur',
-          symbol: 'EUR',
-          name: 'Euro',
-          decimals: 2,
-          denomSymbol: '€',
-        },
-        fiatId: '/currencies/fiat/eur',
-        amountIn: 100,
-        amountOut: 0.035741,
-        exchangeRate: 2797.9071654402505,
-        networkFee: 0,
-        providerFee: 0,
-        extraFee: 0,
-        receiver: '0x76cf1CdD1fcC252442b50D6e97207228aA4aefC3',
-        paymentMethod: '/payments/apple-pay',
-        cryptoTranslation: 'eth-eth',
-      },
-      nativeApplePay: {
-        supported: false,
-        merchantId: 'merchant.io.metamask.banxa.test',
-      },
-      providerInfo: {
-        id: '/providers/banxa-staging',
-        name: 'Banxa (Staging)',
-        environmentType: 'STAGING',
-        description:
-          "Per Banxa: “Established from 2014, Banxa is the world's first publicly listed Financial technology platform, powering a world-leading fiat to crypto gateway solution for customers to buy, sell and trade digital assets. Banxa's payment infrastructure offers online payment services across multiple currencies, crypto, and payment types from card to local bank transfers. Banxa now supports over 130+ countries and more than 80 currencies.”",
-        hqAddress: '2/6 Gwynne St, Cremorne VIC 3121, Australia',
-        links: [
-          {
-            name: 'Homepage',
-            url: 'https://banxa.com/',
-          },
-          {
-            name: 'Terms of Service',
-            url: 'https://banxa.com/wp-content/uploads/2022/10/Customer-Terms-and-Conditions-1-July-2022.pdf',
-          },
-          {
-            name: 'Support',
-            url: 'https://support.banxa.com/en/support/tickets/new',
-          },
-        ],
-        logos: {
-          light: '/assets/providers/banxa_light.png',
-          dark: '/assets/providers/banxa_dark.png',
-          height: 24,
-          width: 65,
-        },
-        features: {
-          buy: {
-            userAgent: null,
-            padCustomOrderId: false,
-            orderCustomId: '',
-            browser: 'APP_BROWSER',
-            orderCustomIdRequired: false,
-            orderCustomIdExpiration: null,
-            orderCustomIdSeparator: null,
-            orderCustomIdPrefixes: ['c-', ''],
-            supportedByBackend: true,
-            redirection: 'JSON_REDIRECTION',
-          },
-          quotes: {
-            enabled: false,
-            supportedByBackend: false,
-          },
-          sell: {
-            enabled: true,
-          },
-          sellQuotes: {
-            enabled: true,
-          },
-          recurringBuy: {},
-        },
-      },
-      providerQuote: {
-        paymentMethod: ['apple-pay'],
-        fiatCurrency: 'EUR',
-        cryptoCurrency: 'eth-eth',
-      },
-      metadata: {
-        reliability: 0.6118746564,
-        tags: {
-          isBestRate: true,
-        },
-      },
-    },
-    {
-      provider: '/providers/banxa',
-      url: null,
-      method: null,
-      headers: null,
-      quote: {
-        crypto: {
-          id: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          idv2: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          legacyId: '/currencies/crypto/1/eth',
-          network: {
-            active: true,
-            chainId: '1',
-            chainName: 'Ethereum Mainnet',
-            shortName: 'Ethereum',
-          },
-          logo: 'https://uat-static.cx.metamask.io/api/v1/tokenIcons/1/0x0000000000000000000000000000000000000000.png',
-          decimals: 18,
-          address: '0x0000000000000000000000000000000000000000',
-          symbol: 'ETH',
-          name: 'Ethereum',
-        },
-        cryptoId:
-          '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-        fiat: {
-          id: '/currencies/fiat/eur',
-          symbol: 'EUR',
-          name: 'Euro',
-          decimals: 2,
-          denomSymbol: '€',
-        },
-        fiatId: '/currencies/fiat/eur',
-        amountIn: 100,
-        amountOut: 0.024672,
-        exchangeRate: 3972.924773022049,
-        networkFee: 0.03,
-        providerFee: 1.95,
-        extraFee: 0,
-        receiver: '0x76cf1CdD1fcC252442b50D6e97207228aA4aefC3',
-        paymentMethod: '/payments/apple-pay',
-        cryptoTranslation: 'eth-eth',
-      },
-      nativeApplePay: {
-        supported: false,
-        merchantId: 'merchant.io.metamask.banxa',
-      },
-      providerInfo: {
-        id: '/providers/banxa',
-        name: 'Banxa',
-        environmentType: 'PRODUCTION',
-        description:
-          "Per Banxa: “Established from 2014, Banxa is the world's first publicly listed Financial technology platform, powering a world-leading fiat to crypto gateway solution for customers to buy, sell and trade digital assets. Banxa's payment infrastructure offers online payment services across multiple currencies, crypto, and payment types from card to local bank transfers. Banxa now supports over 130+ countries and more than 80 currencies.”",
-        hqAddress: '2/6 Gwynne St, Cremorne VIC 3121, Australia',
-        links: [
-          {
-            name: 'Homepage',
-            url: 'https://banxa.com/',
-          },
-          {
-            name: 'Terms of Service',
-            url: 'https://banxa.com/wp-content/uploads/2022/10/Customer-Terms-and-Conditions-1-July-2022.pdf',
-          },
-          {
-            name: 'Support',
-            url: 'https://support.banxa.com/en/support/tickets/new',
-          },
-        ],
-        logos: {
-          light: '/assets/providers/banxa_light.png',
-          dark: '/assets/providers/banxa_dark.png',
-          height: 24,
-          width: 65,
-        },
-        features: {
-          buy: {
-            userAgent: null,
-            padCustomOrderId: false,
-            orderCustomId: '',
-            browser: 'APP_BROWSER',
-            orderCustomIdRequired: false,
-            orderCustomIdExpiration: null,
-            orderCustomIdSeparator: null,
-            orderCustomIdPrefixes: ['c-', ''],
-            supportedByBackend: true,
-            redirection: 'JSON_REDIRECTION',
-          },
-          quotes: {
-            enabled: false,
-            supportedByBackend: false,
-          },
-          sell: {
-            enabled: true,
-          },
-          sellQuotes: {
-            enabled: true,
-          },
-          recurringBuy: {},
-        },
-      },
-      providerQuote: {
-        paymentMethod: ['apple-pay'],
-        fiatCurrency: 'EUR',
-        cryptoCurrency: 'eth-eth',
-      },
-      metadata: {
-        reliability: 0.6118746564,
-        tags: {},
-      },
-    },
-    {
-      provider: '/providers/revolut-staging',
-      url: null,
-      method: null,
-      headers: null,
-      quote: {
-        crypto: {
-          id: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          idv2: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          legacyId: '/currencies/crypto/1/eth',
-          network: {
-            active: true,
-            chainId: '1',
-            chainName: 'Ethereum Mainnet',
-            shortName: 'Ethereum',
-          },
-          logo: 'https://uat-static.cx.metamask.io/api/v1/tokenIcons/1/0x0000000000000000000000000000000000000000.png',
-          decimals: 18,
-          address: '0x0000000000000000000000000000000000000000',
-          symbol: 'ETH',
-          name: 'Ethereum',
-        },
-        cryptoId:
-          '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-        fiat: {
-          id: '/currencies/fiat/eur',
-          symbol: 'EUR',
-          name: 'Euro',
-          decimals: 2,
-          denomSymbol: '€',
-        },
-        fiatId: '/currencies/fiat/eur',
-        amountIn: 100,
-        amountOut: 0.02427084,
-        exchangeRate: 4015.930227383972,
-        networkFee: 0.09,
-        providerFee: 2.44,
-        extraFee: 0,
-        receiver: '0x76cf1CdD1fcC252442b50D6e97207228aA4aefC3',
-        paymentMethod: '/payments/apple-pay',
-        cryptoTranslation: 'ETH',
-      },
-      nativeApplePay: {},
-      providerInfo: {
-        id: '/providers/revolut-staging',
-        name: 'Revolut (Staging)',
-        environmentType: 'STAGING',
-        description:
-          'Per Revolut: "Revolut is a global super-app used by over 30M+ people to manage all things money. Existing Revolut customers can complete a transaction in a few taps (no KYC required), while having the option to pay directly from their Revolut account with lower fees. If you don\'t have a Revolut account, enjoy a £20 cashback welcome offer when you set up a Revolut account after you make a successful Metamask transaction via Revolut."',
-        hqAddress: '7 Westferry Circus, London E14 4HD, United Kingdom',
-        links: [
-          {
-            name: 'Homepage',
-            url: 'https://www.revolut.com',
-          },
-          {
-            name: 'Terms of Service',
-            url: 'https://www.revolut.com/legal/revramp',
-          },
-          {
-            name: 'Support',
-            url: 'https://ramp.revolut.com/en/support',
-          },
-        ],
-        logos: {
-          light: '/assets/providers/revolut_light.png',
-          dark: '/assets/providers/revolut_dark.png',
-          height: 24,
-          width: 58,
-        },
-        features: {
-          buy: {
-            enabled: true,
-            userAgent: null,
-            padCustomOrderId: true,
-            orderCustomId: 'GUID',
-            browser: 'APP_BROWSER',
-            orderCustomIdRequired: true,
-            orderCustomIdExpiration: 60,
-            orderCustomIdSeparator: null,
-            orderCustomIdPrefixes: [],
-            supportedByBackend: true,
-            redirection: 'JSON_REDIRECTION',
-          },
-          quotes: {
-            enabled: false,
-            supportedByBackend: false,
-          },
-          sell: {
-            enabled: true,
-          },
-          sellQuotes: {
-            enabled: true,
-          },
-          recurringBuy: {},
-        },
-      },
-      providerQuote: {
-        paymentMethod: ['apple-pay'],
-        fiatCurrency: 'EUR',
-        cryptoCurrency: 'ETH',
-      },
-      metadata: {
-        reliability: 0.5063418563,
-        tags: {},
-      },
-    },
-    {
-      provider: '/providers/revolut',
-      url: null,
-      method: null,
-      headers: null,
-      quote: {
-        crypto: {
-          id: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          idv2: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          legacyId: '/currencies/crypto/1/eth',
-          network: {
-            active: true,
-            chainId: '1',
-            chainName: 'Ethereum Mainnet',
-            shortName: 'Ethereum',
-          },
-          logo: 'https://uat-static.cx.metamask.io/api/v1/tokenIcons/1/0x0000000000000000000000000000000000000000.png',
-          decimals: 18,
-          address: '0x0000000000000000000000000000000000000000',
-          symbol: 'ETH',
-          name: 'Ethereum',
-        },
-        cryptoId:
-          '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-        fiat: {
-          id: '/currencies/fiat/eur',
-          symbol: 'EUR',
-          name: 'Euro',
-          decimals: 2,
-          denomSymbol: '€',
-        },
-        fiatId: '/currencies/fiat/eur',
-        amountIn: 100,
-        amountOut: 0.02416794,
-        exchangeRate: 4035.5115082212224,
-        networkFee: 0.03,
-        providerFee: 2.44,
-        extraFee: 0,
-        receiver: '0x76cf1CdD1fcC252442b50D6e97207228aA4aefC3',
-        paymentMethod: '/payments/apple-pay',
-        cryptoTranslation: 'ETH',
-      },
-      nativeApplePay: {},
-      providerInfo: {
-        id: '/providers/revolut',
-        name: 'Revolut',
-        environmentType: 'PRODUCTION',
-        description:
-          'Per Revolut: "Revolut is a global super-app used by over 30M+ people to manage all things money. Existing Revolut customers can complete a transaction in a few taps (no KYC required), while having the option to pay directly from their Revolut account with lower fees. If you don\'t have a Revolut account, enjoy a £20 cashback welcome offer when you set up a Revolut account after you make a successful Metamask transaction via Revolut."',
-        hqAddress: '7 Westferry Circus, London E14 4HD, United Kingdom',
-        links: [
-          {
-            name: 'Homepage',
-            url: 'https://www.revolut.com',
-          },
-          {
-            name: 'Terms of Service',
-            url: 'https://www.revolut.com/legal/revramp',
-          },
-          {
-            name: 'Support',
-            url: 'https://ramp.revolut.com/en/support',
-          },
-        ],
-        logos: {
-          light: '/assets/providers/revolut_light.png',
-          dark: '/assets/providers/revolut_dark.png',
-          height: 24,
-          width: 58,
-        },
-        features: {
-          buy: {
-            enabled: true,
-            userAgent: null,
-            padCustomOrderId: true,
-            orderCustomId: 'GUID',
-            browser: 'APP_BROWSER',
-            orderCustomIdRequired: true,
-            orderCustomIdExpiration: 60,
-            orderCustomIdSeparator: null,
-            orderCustomIdPrefixes: [],
-            supportedByBackend: true,
-            redirection: 'JSON_REDIRECTION',
-          },
-          quotes: {
-            enabled: false,
-            supportedByBackend: false,
-          },
-          sell: {
-            enabled: true,
-          },
-          sellQuotes: {
-            enabled: true,
-          },
-          recurringBuy: {},
-        },
-      },
-      providerQuote: {
-        paymentMethod: ['apple-pay'],
-        fiatCurrency: 'EUR',
-        cryptoCurrency: 'ETH',
-      },
-      metadata: {
-        reliability: 0.5063418563,
-        tags: {},
-      },
-    },
-    {
-      provider: '/providers/ramp-network',
-      url: null,
-      method: null,
-      headers: null,
-      quote: {
-        crypto: {
-          id: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          idv2: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          legacyId: '/currencies/crypto/1/eth',
-          network: {
-            active: true,
-            chainId: '1',
-            chainName: 'Ethereum Mainnet',
-            shortName: 'Ethereum',
-          },
-          logo: 'https://uat-static.cx.metamask.io/api/v1/tokenIcons/1/0x0000000000000000000000000000000000000000.png',
-          decimals: 18,
-          address: '0x0000000000000000000000000000000000000000',
-          symbol: 'ETH',
-          name: 'Ethereum',
-        },
-        cryptoId:
-          '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-        fiat: {
-          id: '/currencies/fiat/eur',
-          symbol: 'EUR',
-          name: 'Euro',
-          decimals: 2,
-          denomSymbol: '€',
-        },
-        fiatId: '/currencies/fiat/eur',
-        amountIn: 100,
-        amountOut: 0.023741401691137465,
-        exchangeRate: 4027.984583391222,
-        networkFee: 0.07,
-        providerFee: 4.3,
-        extraFee: 0,
-        receiver: '0x76cf1CdD1fcC252442b50D6e97207228aA4aefC3',
-        paymentMethod: '/payments/apple-pay',
-        cryptoTranslation: 'ETH_ETH',
-      },
-      nativeApplePay: {},
-      providerInfo: {
-        id: '/providers/ramp-network',
-        name: 'Ramp Network',
-        environmentType: 'PRODUCTION',
-        description:
-          'Ramp is building the infrastructure to seamlessly connect Web3 with today’s global financial system. Through its core on- and off-ramp products, Ramp provides businesses and individuals across 150+ countries with a streamlined and smooth experience in converting between crypto and fiat currencies. Ramp is fully integrated with the world’s major payment methods, including debit and credit cards, bank transfers, Apple Pay, Google Pay, and more.',
-        hqAddress: '8 The Green, STE B, Dover, County of Kent, DE 19901, USA',
-        links: [
-          {
-            name: 'Homepage',
-            url: 'https://ramp.network/',
-          },
-          {
-            name: 'Terms of Service',
-            url: 'https://ramp.network/terms-of-service',
-          },
-          {
-            name: 'Support',
-            url: 'https://support.ramp.network/en/collections/6690-customer-support-help-center',
-          },
-        ],
-        logos: {
-          light: '/assets/providers/ramp-logo-light.png',
-          dark: '/assets/providers/ramp-logo-dark.png',
-          height: 24,
-          width: 79,
-        },
-        features: {
-          buy: {
-            enabled: true,
-            userAgent: null,
-            padCustomOrderId: true,
-            orderCustomId: 'DIGITS_10',
-            browser: 'APP_BROWSER',
-            orderCustomIdRequired: true,
-            orderCustomIdExpiration: 60,
-            orderCustomIdSeparator: null,
-            orderCustomIdPrefixes: [],
-            supportedByBackend: true,
-            redirection: 'JSON_REDIRECTION',
-          },
-          quotes: {
-            enabled: false,
-            supportedByBackend: false,
-          },
-          sell: {
-            enabled: true,
-          },
-          sellQuotes: {
-            enabled: true,
-          },
-          recurringBuy: {},
-        },
-      },
-      providerQuote: {
-        paymentMethod: ['apple-pay'],
-        fiatCurrency: 'EUR',
-        cryptoCurrency: 'ETH_ETH',
-      },
-      metadata: {
-        reliability: 0.5046052632,
-        tags: {},
-      },
-    },
-    {
-      provider: '/providers/ramp-network-staging',
-      url: null,
-      method: null,
-      headers: null,
-      quote: {
-        crypto: {
-          id: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          idv2: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          legacyId: '/currencies/crypto/1/eth',
-          network: {
-            active: true,
-            chainId: '1',
-            chainName: 'Ethereum Mainnet',
-            shortName: 'Ethereum',
-          },
-          logo: 'https://uat-static.cx.metamask.io/api/v1/tokenIcons/1/0x0000000000000000000000000000000000000000.png',
-          decimals: 18,
-          address: '0x0000000000000000000000000000000000000000',
-          symbol: 'ETH',
-          name: 'Ethereum',
-        },
-        cryptoId:
-          '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-        fiat: {
-          id: '/currencies/fiat/eur',
-          symbol: 'EUR',
-          name: 'Euro',
-          decimals: 2,
-          denomSymbol: '€',
-        },
-        fiatId: '/currencies/fiat/eur',
-        amountIn: 100,
-        amountOut: 0.023486763146531114,
-        exchangeRate: 4028.2264273599344,
-        networkFee: 1.7,
-        providerFee: 3.69,
-        extraFee: 0,
-        receiver: '0x76cf1CdD1fcC252442b50D6e97207228aA4aefC3',
-        paymentMethod: '/payments/apple-pay',
-        cryptoTranslation: 'SEPOLIA_ETH',
-      },
-      nativeApplePay: {},
-      providerInfo: {
-        id: '/providers/ramp-network-staging',
-        name: 'Ramp Network (Staging)',
-        environmentType: 'STAGING',
-        description:
-          'Ramp is building the infrastructure to seamlessly connect Web3 with today’s global financial system. Through its core on- and off-ramp products, Ramp provides businesses and individuals across 150+ countries with a streamlined and smooth experience in converting between crypto and fiat currencies. Ramp is fully integrated with the world’s major payment methods, including debit and credit cards, bank transfers, Apple Pay, Google Pay, and more.',
-        hqAddress: '8 The Green, STE B, Dover, County of Kent, DE 19901, USA',
-        links: [
-          {
-            name: 'Homepage',
-            url: 'https://ramp.network/',
-          },
-          {
-            name: 'Terms of Service',
-            url: 'https://ramp.network/terms-of-service',
-          },
-          {
-            name: 'Support',
-            url: 'https://support.ramp.network/en/collections/6690-customer-support-help-center',
-          },
-        ],
-        logos: {
-          light: '/assets/providers/ramp-logo-light.png',
-          dark: '/assets/providers/ramp-logo-dark.png',
-          height: 24,
-          width: 79,
-        },
-        features: {
-          buy: {
-            enabled: true,
-            userAgent: null,
-            padCustomOrderId: true,
-            orderCustomId: 'DIGITS_10',
-            browser: 'APP_BROWSER',
-            orderCustomIdRequired: true,
-            orderCustomIdExpiration: 60,
-            orderCustomIdSeparator: null,
-            orderCustomIdPrefixes: [],
-            supportedByBackend: true,
-            redirection: 'JSON_REDIRECTION',
-          },
-          quotes: {
-            enabled: false,
-            supportedByBackend: false,
-          },
-          sell: {
-            enabled: true,
-          },
-          sellQuotes: {
-            enabled: true,
-          },
-          recurringBuy: {},
-        },
-      },
-      providerQuote: {
-        paymentMethod: ['apple-pay'],
-        fiatCurrency: 'EUR',
-        cryptoCurrency: 'SEPOLIA_ETH',
-      },
-      metadata: {
-        reliability: 0.5046052632,
-        tags: {},
-      },
-    },
-    {
-      provider: '/providers/sardine-staging',
-      url: null,
-      method: null,
-      headers: null,
-      quote: {
-        crypto: {
-          id: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          idv2: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          legacyId: '/currencies/crypto/1/eth',
-          network: {
-            active: true,
-            chainId: '1',
-            chainName: 'Ethereum Mainnet',
-            shortName: 'Ethereum',
-          },
-          logo: 'https://uat-static.cx.metamask.io/api/v1/tokenIcons/1/0x0000000000000000000000000000000000000000.png',
-          decimals: 18,
-          address: '0x0000000000000000000000000000000000000000',
-          symbol: 'ETH',
-          name: 'Ethereum',
-        },
-        cryptoId:
-          '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-        fiat: {
-          id: '/currencies/fiat/eur',
-          symbol: 'EUR',
-          name: 'Euro',
-          decimals: 2,
-          denomSymbol: '€',
-        },
-        fiatId: '/currencies/fiat/eur',
-        amountIn: 100,
-        amountOut: 0.025497793837794054,
-        exchangeRate: 3920.9044196526966,
-        networkFee: 0.02558745,
-        providerFee: 0,
-        extraFee: 0,
-        receiver: '0x76cf1CdD1fcC252442b50D6e97207228aA4aefC3',
-        paymentMethod: '/payments/apple-pay',
-        cryptoTranslation: 'ETH:ethereum',
-      },
-      nativeApplePay: {},
-      providerInfo: {
-        id: '/providers/sardine-staging',
-        name: 'Sardine (Staging)',
-        environmentType: 'STAGING',
-        description:
-          'Per Sardine: “Sardine provides the safest, most convenient way to purchase cryptocurrencies and convert between fiat currencies. We offer lower transaction fees, higher transaction limits, instant account funding, and support for all major payment methods across 180+ countries."',
-        hqAddress: '82 NE 191st St, #58243 Miami, Florida 33179-3899 US',
-        links: [
-          {
-            name: 'Homepage',
-            url: 'https://www.sardine.ai/',
-          },
-          {
-            name: 'Privacy Policy',
-            url: 'https://crypto.sardine.ai/privacy',
-          },
-          {
-            name: 'Support',
-            url: 'https://crypto.sardine.ai/support',
-          },
-        ],
-        logos: {
-          light: '/assets/providers/sardine_light.png',
-          dark: '/assets/providers/sardine_dark.png',
-          height: 24,
-          width: 79,
-        },
-        features: {
-          buy: {
-            enabled: true,
-            userAgent: null,
-            padCustomOrderId: false,
-            orderCustomId: '',
-            browser: 'APP_BROWSER',
-            orderCustomIdRequired: false,
-            orderCustomIdExpiration: null,
-            orderCustomIdSeparator: null,
-            orderCustomIdPrefixes: ['c-', ''],
-            supportedByBackend: true,
-            redirection: 'JSON_REDIRECTION',
-          },
-          quotes: {
-            enabled: false,
-            supportedByBackend: false,
-          },
-          sell: {
-            enabled: true,
-          },
-          sellQuotes: {
-            enabled: true,
-          },
-          recurringBuy: {
-            enabled: false,
-          },
-        },
-      },
-      providerQuote: {
-        paymentMethod: ['apple-pay'],
-        fiatCurrency: 'EUR',
-        cryptoCurrency: 'ETH:ethereum',
-      },
-      metadata: {
-        reliability: 0,
-        tags: {},
-      },
-    },
-    {
-      provider: '/providers/sardine',
-      url: null,
-      method: null,
-      headers: null,
-      quote: {
-        crypto: {
-          id: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          idv2: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          legacyId: '/currencies/crypto/1/eth',
-          network: {
-            active: true,
-            chainId: '1',
-            chainName: 'Ethereum Mainnet',
-            shortName: 'Ethereum',
-          },
-          logo: 'https://uat-static.cx.metamask.io/api/v1/tokenIcons/1/0x0000000000000000000000000000000000000000.png',
-          decimals: 18,
-          address: '0x0000000000000000000000000000000000000000',
-          symbol: 'ETH',
-          name: 'Ethereum',
-        },
-        cryptoId:
-          '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-        fiat: {
-          id: '/currencies/fiat/eur',
-          symbol: 'EUR',
-          name: 'Euro',
-          decimals: 2,
-          denomSymbol: '€',
-        },
-        fiatId: '/currencies/fiat/eur',
-        amountIn: 100,
-        amountOut: 0.024538150261037586,
-        exchangeRate: 4074.2440439262605,
-        networkFee: 0.02558745,
-        providerFee: 0,
-        extraFee: 0,
-        receiver: '0x76cf1CdD1fcC252442b50D6e97207228aA4aefC3',
-        paymentMethod: '/payments/apple-pay',
-        cryptoTranslation: 'ETH:ethereum',
-      },
-      nativeApplePay: {},
-      providerInfo: {
-        id: '/providers/sardine',
-        name: 'Sardine',
-        environmentType: 'PRODUCTION',
-        description:
-          'Per Sardine: “Sardine provides the safest, most convenient way to purchase cryptocurrencies and convert between fiat currencies. We offer lower transaction fees, higher transaction limits, instant account funding, and support for all major payment methods across 180+ countries."',
-        hqAddress: '82 NE 191st St, #58243 Miami, Florida 33179-3899 US',
-        links: [
-          {
-            name: 'Homepage',
-            url: 'https://www.sardine.ai/',
-          },
-          {
-            name: 'Privacy Policy',
-            url: 'https://crypto.sardine.ai/privacy',
-          },
-          {
-            name: 'Support',
-            url: 'https://crypto.sardine.ai/support',
-          },
-        ],
-        logos: {
-          light: '/assets/providers/sardine_light.png',
-          dark: '/assets/providers/sardine_dark.png',
-          height: 24,
-          width: 79,
-        },
-        features: {
-          buy: {
-            enabled: true,
-            userAgent: null,
-            padCustomOrderId: false,
-            orderCustomId: '',
-            browser: 'APP_BROWSER',
-            orderCustomIdRequired: false,
-            orderCustomIdExpiration: null,
-            orderCustomIdSeparator: null,
-            orderCustomIdPrefixes: ['c-', ''],
-            supportedByBackend: true,
-            redirection: 'JSON_REDIRECTION',
-          },
-          quotes: {
-            enabled: false,
-            supportedByBackend: false,
-          },
-          sell: {
-            enabled: true,
-          },
-          sellQuotes: {
-            enabled: true,
-          },
-          recurringBuy: {
-            enabled: false,
-          },
-        },
-      },
-      providerQuote: {
-        paymentMethod: ['apple-pay'],
-        fiatCurrency: 'EUR',
-        cryptoCurrency: 'ETH:ethereum',
-      },
-      metadata: {
-        reliability: 0,
-        tags: {},
-      },
-    },
-    {
-      provider: '/providers/binanceconnect-staging',
-      url: null,
-      method: null,
-      headers: null,
-      quote: {
-        crypto: {
-          id: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          idv2: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          legacyId: '/currencies/crypto/1/eth',
-          network: {
-            active: true,
-            chainId: '1',
-            chainName: 'Ethereum Mainnet',
-            shortName: 'Ethereum',
-          },
-          logo: 'https://uat-static.cx.metamask.io/api/v1/tokenIcons/1/0x0000000000000000000000000000000000000000.png',
-          decimals: 18,
-          address: '0x0000000000000000000000000000000000000000',
-          symbol: 'ETH',
-          name: 'Ethereum',
-        },
-        cryptoId:
-          '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-        fiat: {
-          id: '/currencies/fiat/eur',
-          symbol: 'EUR',
-          name: 'Euro',
-          decimals: 2,
-          denomSymbol: '€',
-        },
-        fiatId: '/currencies/fiat/eur',
-        amountIn: 100,
-        amountOut: 0.02372126,
-        exchangeRate: 4079.0668576714306,
-        networkFee: 1.239394511793,
-        providerFee: 2,
-        extraFee: 0,
-        receiver: '0x76cf1CdD1fcC252442b50D6e97207228aA4aefC3',
-        paymentMethod: '/payments/apple-pay',
-        cryptoTranslation: 'ETH',
-      },
-      nativeApplePay: {},
-      providerInfo: {
-        id: '/providers/binanceconnect-staging',
-        name: 'Binance Connect (Staging)',
-        environmentType: 'STAGING',
-        description:
-          'Binance Connect offers an easy way to buy and sell crypto through your DeFi wallet. You can conveniently purchase crypto using various payment methods, including credit cards, bank transfers, Binance wallet balance, and alternative methods through Binance Peer-to-Peer (P2P) services.',
-        hqAddress: 'House of Francis, Room 303, IIe Du Port, Mahe, Seychelles',
-        links: [
-          {
-            name: 'Homepage',
-            url: 'https://www.binance.com/en/binance-connect/',
-          },
-          {
-            name: 'Terms of Service',
-            url: 'https://www.binance.com/en/terms/',
-          },
-          {
-            name: 'Support',
-            url: 'https://www.binance.com/en/support/',
-          },
-        ],
-        logos: {
-          light: '/assets/providers/binance-connect-light.png',
-          dark: '/assets/providers/binance-connect-dark.png',
-          height: 24,
-          width: 92,
-        },
-        features: {
-          buy: {
-            enabled: true,
-            userAgent: null,
-            padCustomOrderId: false,
-            orderCustomId: '',
-            browser: 'IN_APP_OS_BROWSER',
-            orderCustomIdRequired: false,
-            orderCustomIdExpiration: null,
-            orderCustomIdSeparator: null,
-            orderCustomIdPrefixes: ['c-', ''],
-            supportedByBackend: true,
-            redirection: 'JSON_REDIRECTION',
-          },
-          quotes: {
-            enabled: false,
-            supportedByBackend: false,
-          },
-          sell: {},
-          sellQuotes: {},
-          recurringBuy: {},
-        },
-      },
-      providerQuote: {
-        paymentMethod: ['apple-pay'],
-        fiatCurrency: 'EUR',
-        cryptoCurrency: 'ETH',
-      },
-      metadata: {
-        reliability: 0,
-        tags: {},
-      },
-    },
-    {
-      provider: '/providers/binanceconnect',
-      url: null,
-      method: null,
-      headers: null,
-      quote: {
-        crypto: {
-          id: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          idv2: '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-          legacyId: '/currencies/crypto/1/eth',
-          network: {
-            active: true,
-            chainId: '1',
-            chainName: 'Ethereum Mainnet',
-            shortName: 'Ethereum',
-          },
-          logo: 'https://uat-static.cx.metamask.io/api/v1/tokenIcons/1/0x0000000000000000000000000000000000000000.png',
-          decimals: 18,
-          address: '0x0000000000000000000000000000000000000000',
-          symbol: 'ETH',
-          name: 'Ethereum',
-        },
-        cryptoId:
-          '/currencies/crypto/1/0x0000000000000000000000000000000000000000',
-        fiat: {
-          id: '/currencies/fiat/eur',
-          symbol: 'EUR',
-          name: 'Euro',
-          decimals: 2,
-          denomSymbol: '€',
-        },
-        fiatId: '/currencies/fiat/eur',
-        amountIn: 100,
-        amountOut: 0.02372126,
-        exchangeRate: 4079.0668576714306,
-        networkFee: 1.239394511793,
-        providerFee: 2,
-        extraFee: 0,
-        receiver: '0x76cf1CdD1fcC252442b50D6e97207228aA4aefC3',
-        paymentMethod: '/payments/apple-pay',
-        cryptoTranslation: 'ETH',
-      },
-      nativeApplePay: {},
-      providerInfo: {
-        id: '/providers/binanceconnect',
-        name: 'Binance Connect',
-        environmentType: 'PRODUCTION',
-        description:
-          'Binance Connect offers an easy way to buy and sell crypto through your DeFi wallet. You can conveniently purchase crypto using various payment methods, including credit cards, bank transfers, Binance wallet balance, and alternative methods through Binance Peer-to-Peer (P2P) services.',
-        hqAddress: 'House of Francis, Room 303, IIe Du Port, Mahe, Seychelles',
-        links: [
-          {
-            name: 'Homepage',
-            url: 'https://www.binance.com/en/binance-connect/',
-          },
-          {
-            name: 'Terms of Service',
-            url: 'https://www.binance.com/en/terms/',
-          },
-          {
-            name: 'Support',
-            url: 'https://www.binance.com/en/support/',
-          },
-        ],
-        logos: {
-          light: '/assets/providers/binance-connect-light.png',
-          dark: '/assets/providers/binance-connect-dark.png',
-          height: 24,
-          width: 92,
-        },
-        features: {
-          buy: {
-            enabled: true,
-            userAgent: null,
-            padCustomOrderId: false,
-            orderCustomId: '',
-            browser: 'IN_APP_OS_BROWSER',
-            orderCustomIdRequired: false,
-            orderCustomIdExpiration: null,
-            orderCustomIdSeparator: null,
-            orderCustomIdPrefixes: ['c-', ''],
-            supportedByBackend: true,
-            redirection: 'JSON_REDIRECTION',
-          },
-          quotes: {
-            enabled: false,
-            supportedByBackend: false,
-          },
-          sell: {},
-          sellQuotes: {},
-          recurringBuy: {},
-        },
-      },
-      providerQuote: {
-        paymentMethod: ['apple-pay'],
-        fiatCurrency: 'EUR',
-        cryptoCurrency: 'ETH',
-      },
-      metadata: {
-        reliability: 0,
-        tags: {},
-      },
-    },
-  ],
-  sorted: [
-    {
-      sortBy: '2',
-      ids: [
-        '/providers/banxa-staging',
-        '/providers/sardine-staging',
-        '/providers/moonpay-staging',
-        '/providers/moonpay',
-        '/providers/banxa',
-        '/providers/mercuryo',
-        '/providers/sardine',
-        '/providers/revolut-staging',
-        '/providers/revolut',
-        '/providers/transak',
-        '/providers/ramp-network',
-        '/providers/binanceconnect-staging',
-        '/providers/binanceconnect',
-        '/providers/transak-staging',
-        '/providers/ramp-network-staging',
-        '/providers/mercuryo-staging',
-      ],
-    },
-    {
-      sortBy: '1',
-      ids: [
-        '/providers/mercuryo',
-        '/providers/mercuryo-staging',
-        '/providers/moonpay-staging',
-        '/providers/moonpay',
-        '/providers/transak',
-        '/providers/transak-staging',
-        '/providers/banxa-staging',
-        '/providers/banxa',
-        '/providers/revolut-staging',
-        '/providers/revolut',
-        '/providers/ramp-network',
-        '/providers/ramp-network-staging',
-        '/providers/sardine-staging',
-        '/providers/sardine',
-        '/providers/binanceconnect-staging',
-        '/providers/binanceconnect',
-      ],
-    },
-  ],
-  error: [
-    {
-      providerInfo: {
-        id: '/providers/coinbase-staging',
-        name: 'Coinbase (Staging)',
-        environmentType: 'STAGING',
-        description:
-          'Per Coinbase: "Coinbase Pay enables purchases and transfers directly from your Coinbase account to your Metamask wallet. It integrates with Dapps and self-custody wallets to use your saved payment methods or crypto balances."',
-        hqAddress: '248 3rd St #434, Oakland CA, 94607',
-        links: [
-          {
-            name: 'Homepage',
-            url: 'https://www.coinbase.com/',
-          },
-          {
-            name: 'Terms of Service',
-            url: 'https://www.coinbase.com/legal/',
-          },
-        ],
-        logos: {
-          light: '/assets/providers/coinbase-light.png',
-          dark: '/assets/providers/coinbase-dark.png',
-          height: 24,
-          width: 110,
-        },
-        features: {
-          buy: {
-            enabled: true,
-            userAgent: null,
-            padCustomOrderId: false,
-            orderCustomId: 'GUID',
-            browser: 'IN_APP_OS_BROWSER',
-            orderCustomIdRequired: true,
-            orderCustomIdExpiration: 60,
-            orderCustomIdSeparator: '-',
-            orderCustomIdPrefixes: ['c-', ''],
-            supportedByBackend: true,
-            redirection: 'JSON_REDIRECTION',
-          },
-          quotes: {
-            enabled: false,
-            supportedByBackend: false,
-          },
-          sell: {
-            enabled: false,
-          },
-          sellQuotes: {
-            enabled: false,
-          },
-          recurringBuy: {},
-        },
-      },
-      provider: '/providers/coinbase-staging',
-    },
-    {
-      providerInfo: {
-        id: '/providers/unlimitmeld-staging',
-        name: 'Unlimit (Staging)',
-        environmentType: 'STAGING',
-        description:
-          'Per Unlimit: "Unlimit’s fiat onramp allows users to use a variety of local payment methods to purchase crypto with the lowest fees and highest convenience."',
-        hqAddress: 'Vilnius, Eišiškių Sodų 18-oji g. 11',
-        links: [
-          {
-            name: 'Homepage',
-            url: 'https://www.crypto.unlimit.com',
-          },
-          {
-            name: 'Privacy Policy',
-            url: 'https://cdn.gatefi.com/regform/gatefi-user-terms.pdf',
-          },
-          {
-            name: 'Support',
-            url: 'https://www.crypto.unlimit.com/support/',
-          },
-        ],
-        logos: {
-          light: '/assets/providers/unlimit-light.png',
-          dark: '/assets/providers/unlimit-dark.png',
-          height: 24,
-          width: 58,
-        },
-        features: {
-          buy: {
-            enabled: true,
-            userAgent: null,
-            padCustomOrderId: false,
-            orderCustomId: '',
-            browser: 'APP_BROWSER',
-            orderCustomIdRequired: false,
-            orderCustomIdExpiration: null,
-            orderCustomIdSeparator: null,
-            orderCustomIdPrefixes: ['c-', ''],
-            supportedByBackend: true,
-            redirection: 'JSON_REDIRECTION',
-          },
-          quotes: {
-            enabled: false,
-            supportedByBackend: false,
-          },
-          sell: {
-            enabled: true,
-          },
-          sellQuotes: {
-            enabled: true,
-          },
-          recurringBuy: {},
-        },
-      },
-      provider: '/providers/unlimitmeld-staging',
-    },
-    {
-      providerInfo: {
-        id: '/providers/unlimitmeld',
-        name: 'Unlimit',
-        environmentType: 'PRODUCTION',
-        description:
-          'Per Unlimit: "Unlimit’s fiat onramp allows users to use a variety of local payment methods to purchase crypto with the lowest fees and highest convenience."',
-        hqAddress: 'Vilnius, Eišiškių Sodų 18-oji g. 11',
-        links: [
-          {
-            name: 'Homepage',
-            url: 'https://www.crypto.unlimit.com',
-          },
-          {
-            name: 'Privacy Policy',
-            url: 'https://cdn.gatefi.com/regform/gatefi-user-terms.pdf',
-          },
-          {
-            name: 'Support',
-            url: 'https://www.crypto.unlimit.com/support/',
-          },
-        ],
-        logos: {
-          light: '/assets/providers/unlimit-light.png',
-          dark: '/assets/providers/unlimit-dark.png',
-          height: 24,
-          width: 58,
-        },
-        features: {
-          buy: {
-            enabled: true,
-            userAgent: null,
-            padCustomOrderId: false,
-            orderCustomId: '',
-            browser: 'APP_BROWSER',
-            orderCustomIdRequired: false,
-            orderCustomIdExpiration: null,
-            orderCustomIdSeparator: null,
-            orderCustomIdPrefixes: ['c-', ''],
-            supportedByBackend: true,
-            redirection: 'JSON_REDIRECTION',
-          },
-          quotes: {
-            enabled: false,
-            supportedByBackend: false,
-          },
-          sell: {
-            enabled: true,
-          },
-          sellQuotes: {
-            enabled: true,
-          },
-          recurringBuy: {},
-        },
-      },
-      provider: '/providers/unlimitmeld',
-    },
-    {
-      providerInfo: {
-        id: '/providers/stripe-staging',
-        name: 'Stripe (Staging)',
-        environmentType: 'STAGING',
-        description:
-          'Per Stripe: "Millions of companies—from the world’s largest enterprises to the most ambitious startups—use Stripe to accept payments and grow their revenue. The Stripe crypto onramp gives web3 developers an easy and fast way for their consumers to purchase crypto with a credit card, debit card, or ACH. Top web3 companies trust Stripe\'s crypto onramp to give users a seamless experience that is optimized for conversion and instant settlement of crypto."',
-        hqAddress: '354 Oyster Point Blvd, South San Francisco, CA 94080, USA',
-        links: [
-          {
-            name: 'Homepage',
-            url: 'https://crypto.link.com',
-          },
-          {
-            name: 'Support',
-            url: 'https://support.link.com/topics/crypto',
-          },
-          {
-            name: 'Terms of Service',
-            url: 'https://link.com/terms/crypto-onramp',
-          },
-          {
-            name: 'Privacy Policy',
-            url: 'https://link.com/privacy',
-          },
-        ],
-        logos: {
-          light: '/assets/providers/stripe_light.png',
-          dark: '/assets/providers/stripe_dark.png',
-          height: 24,
-          width: 100,
-        },
-        features: {
-          buy: {
-            enabled: true,
-            userAgent: null,
-            padCustomOrderId: false,
-            orderCustomId: 'GUID',
-            browser: 'APP_BROWSER',
-            orderCustomIdRequired: true,
-            orderCustomIdExpiration: 60,
-            orderCustomIdSeparator: '-',
-            orderCustomIdPrefixes: ['c-', ''],
-            supportedByBackend: true,
-            redirection: 'JSON_REDIRECTION',
-          },
-          quotes: {
-            enabled: false,
-            supportedByBackend: false,
-          },
-          sell: {},
-          sellQuotes: {},
-          recurringBuy: {},
-        },
-      },
-      provider: '/providers/stripe-staging',
-    },
-    {
-      providerInfo: {
-        id: '/providers/stripe',
-        name: 'Stripe',
-        environmentType: 'PRODUCTION',
-        description:
-          'Per Stripe: "Millions of companies—from the world’s largest enterprises to the most ambitious startups—use Stripe to accept payments and grow their revenue. The Stripe crypto onramp gives web3 developers an easy and fast way for their consumers to purchase crypto with a credit card, debit card, or ACH. Top web3 companies trust Stripe\'s crypto onramp to give users a seamless experience that is optimized for conversion and instant settlement of crypto."',
-        hqAddress: '354 Oyster Point Blvd, South San Francisco, CA 94080, USA',
-        links: [
-          {
-            name: 'Homepage',
-            url: 'https://crypto.link.com',
-          },
-          {
-            name: 'Support',
-            url: 'https://support.link.com/topics/crypto',
-          },
-          {
-            name: 'Terms of Service',
-            url: 'https://link.com/terms/crypto-onramp',
-          },
-          {
-            name: 'Privacy Policy',
-            url: 'https://link.com/privacy',
-          },
-        ],
-        logos: {
-          light: '/assets/providers/stripe_light.png',
-          dark: '/assets/providers/stripe_dark.png',
-          height: 24,
-          width: 100,
-        },
-        features: {
-          buy: {
-            enabled: true,
-            userAgent: null,
-            padCustomOrderId: false,
-            orderCustomId: 'GUID',
-            browser: 'APP_BROWSER',
-            orderCustomIdRequired: true,
-            orderCustomIdExpiration: 60,
-            orderCustomIdSeparator: '-',
-            orderCustomIdPrefixes: ['c-', ''],
-            supportedByBackend: true,
-            redirection: 'JSON_REDIRECTION',
-          },
-          quotes: {
-            enabled: false,
-            supportedByBackend: false,
-          },
-          sell: {},
-          sellQuotes: {},
-          recurringBuy: {},
-        },
-      },
-      provider: '/providers/stripe',
-    },
-    {
-      providerInfo: {
-        id: '/providers/transak-native-staging',
-        name: 'Transak Native (Staging)',
-        environmentType: 'STAGING',
-        description:
-          'Per Transak: "The fastest and securest way to buy 100+ cryptocurrencies on 75+ blockchains. Pay via Apple Pay, UPI, bank transfer or use your debit or credit card. Trusted by 2+ million global users. Transak empowers wallets, gaming, DeFi, NFTs, Exchanges, and DAOs across 125+ countries."',
-        hqAddress:
-          '35 Shearing Street, Bury St. Edmunds, IP32 6FE, United Kingdom',
-        links: [
-          {
-            name: 'Homepage',
-            url: 'https://www.transak.com/',
-          },
-          {
-            name: 'Privacy Policy',
-            url: 'https://transak.com/privacy-policy',
-          },
-          {
-            name: 'Support',
-            url: 'https://support.transak.com/hc/en-us',
-          },
-        ],
-        logos: {
-          light: '/assets/providers/transak_light.png',
-          dark: '/assets/providers/transak_dark.png',
-          height: 24,
-          width: 90,
-        },
-        features: {
-          buy: {
-            enabled: true,
-            userAgent: null,
-            padCustomOrderId: false,
-            orderCustomId: '',
-            browser: null,
-            orderCustomIdRequired: false,
-            orderCustomIdExpiration: null,
-            orderCustomIdSeparator: null,
-            orderCustomIdPrefixes: ['c-', ''],
-            supportedByBackend: true,
-            redirection: 'JSON_REDIRECTION',
-          },
-          quotes: {
-            enabled: false,
-            supportedByBackend: false,
-          },
-          sell: {
-            enabled: true,
-          },
-          sellQuotes: {
-            enabled: true,
-          },
-          recurringBuy: {},
-        },
-      },
-      provider: '/providers/transak-native-staging',
-    },
-    {
-      providerInfo: {
-        id: '/providers/transak-native',
-        name: 'Transak Native',
-        environmentType: 'PRODUCTION',
-        description:
-          'Per Transak: "The fastest and securest way to buy 100+ cryptocurrencies on 75+ blockchains. Pay via Apple Pay, UPI, bank transfer or use your debit or credit card. Trusted by 2+ million global users. Transak empowers wallets, gaming, DeFi, NFTs, Exchanges, and DAOs across 125+ countries."',
-        hqAddress:
-          '35 Shearing Street, Bury St. Edmunds, IP32 6FE, United Kingdom',
-        links: [
-          {
-            name: 'Homepage',
-            url: 'https://www.transak.com/',
-          },
-          {
-            name: 'Privacy Policy',
-            url: 'https://transak.com/privacy-policy',
-          },
-          {
-            name: 'Support',
-            url: 'https://support.transak.com/hc/en-us',
-          },
-        ],
-        logos: {
-          light: '/assets/providers/transak_light.png',
-          dark: '/assets/providers/transak_dark.png',
-          height: 24,
-          width: 90,
-        },
-        features: {
-          buy: {
-            enabled: true,
-            userAgent: null,
-            padCustomOrderId: false,
-            orderCustomId: '',
-            browser: null,
-            orderCustomIdRequired: false,
-            orderCustomIdExpiration: null,
-            orderCustomIdSeparator: null,
-            orderCustomIdPrefixes: ['c-', ''],
-            supportedByBackend: true,
-            redirection: 'JSON_REDIRECTION',
-          },
-          quotes: {
-            enabled: false,
-            supportedByBackend: false,
-          },
-          sell: {
-            enabled: true,
-          },
-          sellQuotes: {
-            enabled: true,
-          },
-          recurringBuy: {},
-        },
-      },
-      provider: '/providers/transak-native',
-    },
-    {
-      providerInfo: {
-        id: '/providers/mockprovider-staging',
-        name: 'MockProvider (Staging)',
-        environmentType: 'STAGING',
-        description:
-          "Per Banxa: “Established from 2014, Banxa is the world's first publicly listed Financial technology platform, powering a world-leading fiat to crypto gateway solution for customers to buy, sell and trade digital assets. Banxa's payment infrastructure offers online payment services across multiple currencies, crypto, and payment types from card to local bank transfers. Banxa now supports over 130+ countries and more than 80 currencies.”",
-        hqAddress: '2/6 Gwynne St, Cremorne VIC 3121, Australia',
-        links: [
-          {
-            name: 'Homepage',
-            url: 'https://banxa.com/',
-          },
-          {
-            name: 'Terms of Service',
-            url: 'https://banxa.com/wp-content/uploads/2022/10/Customer-Terms-and-Conditions-1-July-2022.pdf',
-          },
-          {
-            name: 'Support',
-            url: 'https://support.banxa.com/en/support/tickets/new',
-          },
-        ],
-        logos: {
-          light: '/assets/providers/banxa_light.png',
-          dark: '/assets/providers/banxa_dark.png',
-          height: 24,
-          width: 65,
-        },
-        features: {
-          buy: {
-            enabled: true,
-            userAgent: null,
-            padCustomOrderId: false,
-            orderCustomId: '',
-            browser: null,
-            orderCustomIdRequired: false,
-            orderCustomIdExpiration: null,
-            orderCustomIdSeparator: null,
-            orderCustomIdPrefixes: ['c-', ''],
-            supportedByBackend: true,
-            redirection: 'JSON_REDIRECTION',
-          },
-          quotes: {
-            enabled: false,
-            supportedByBackend: false,
-          },
-          sell: {
-            enabled: false,
-          },
-          sellQuotes: {
-            enabled: false,
-          },
-          recurringBuy: {},
-        },
-      },
-      provider: '/providers/mockprovider-staging',
-    },
-  ],
-  customActions: [],
-};
