--- tests/api-mocking/mock-responses/ramps/ramps-region-aware-mock-setup.ts
@@ -1,159 +1,0 @@
-import { Mockttp } from 'mockttp';
-import { setupMockRequest } from '../../helpers/mockHelpers.ts';
-import { MockApiEndpoint, RampsRegion } from '../../../framework/types.ts';
-import {
-  RAMPS_NETWORKS_RESPONSE,
-  RAMPS_COUNTRIES_RESPONSE,
-  RAMPS_LIGHT_RESPONSE,
-  RAMPS_AMOUNT_RESPONSE,
-  RAMPS_TOP_TOKENS_RESPONSE,
-  RAMPS_PROVIDERS_RESPONSE,
-  RAMPS_PAYMENTS_V2_RESPONSE,
-} from './ramps-mocks.ts';
-import { createGeolocationResponse } from './ramps-geolocation.ts';
-import { RAMPS_QUOTE_RESPONSE } from './ramps-quotes-response.ts';
-
-/**
- * Sets up comprehensive on-ramp API mocks that are aware of the selected region
- * This ensures the geolocation response matches the region set in the test fixture
- *
- * @param mockServer - The mock server instance
- * @param selectedRegion - The region selected in the test fixture
- */
-export const setupRegionAwareOnRampMocks = async (
-  mockServer: Mockttp,
-  selectedRegion: RampsRegion,
-) => {
-  const geolocationResponse = createGeolocationResponse(selectedRegion);
-
-  const mockEndpoints: MockApiEndpoint[] = [
-    // 1. Geolocation endpoints (UAT and dev only) - region-specific
-    ...geolocationResponse,
-
-    // 2. Region eligibility endpoints (UAT only) - CRITICAL for unified routing
-    {
-      urlEndpoint:
-        /^https:\/\/on-ramp-content\.uat-api\.cx\.metamask\.io\/regions\/countries\/[^/]+$/,
-      responseCode: 200,
-      response: {
-        global: true,
-        deposit: true,
-        aggregator: true,
-      },
-    },
-
-    // 3. Networks endpoints (UAT only)
-    {
-      urlEndpoint:
-        /^https:\/\/on-ramp-cache\.uat-api\.cx\.metamask\.io\/regions\/networks\?.*$/,
-      responseCode: 200,
-      response: RAMPS_NETWORKS_RESPONSE,
-    },
-
-    // 4. Countries endpoints (UAT only) - Add more countries as needed - RAMPS_COUNTRIES_RESPONSE in on-ramp-mocks.ts
-    {
-      urlEndpoint:
-        /^https:\/\/on-ramp-cache\.uat-api\.cx\.metamask\.io(\/v2)?\/regions\/countries\?.*$/,
-      responseCode: 200,
-      response: RAMPS_COUNTRIES_RESPONSE,
-    },
-
-    // 5. Light endpoints - all parameter variations (UAT only)
-    // This controls things like: payment methods, available cryptocurrencies, fiat currencies and limits
-    {
-      urlEndpoint:
-        /^https:\/\/on-ramp-cache\.uat-api\.cx\.metamask\.io\/regions\/[^/]+\/light\?.*$/,
-      responseCode: 200,
-      response: RAMPS_LIGHT_RESPONSE,
-    },
-
-    // 6. Amount conversion endpoints (UAT only)
-    {
-      urlEndpoint:
-        /^https:\/\/on-ramp-cache\.uat-api\.cx\.metamask\.io\/currencies\/crypto\/.*\/amount\?.*$/,
-      responseCode: 200,
-      response: RAMPS_AMOUNT_RESPONSE,
-    },
-
-    // 7. Quote endpoints V1 (UAT only)
-    {
-      urlEndpoint:
-        /^https:\/\/on-ramp\.uat-api\.cx\.metamask\.io\/providers\/all\/quote\?.*$/,
-      responseCode: 200,
-      response: RAMPS_QUOTE_RESPONSE,
-    },
-
-    // 8. Quote endpoints V2 - Unified Ramps (UAT only)
-    {
-      urlEndpoint:
-        /^https:\/\/on-ramp\.uat-api\.cx\.metamask\.io\/v2\/quotes\?.*$/,
-      responseCode: 200,
-      response: RAMPS_QUOTE_RESPONSE,
-    },
-
-    // 9. Top Tokens V2 - Unified Buy (UAT only)
-    {
-      urlEndpoint:
-        /^https:\/\/on-ramp-cache\.uat-api\.cx\.metamask\.io\/v2\/regions\/[^/]+\/topTokens\?.*$/,
-      responseCode: 200,
-      response: RAMPS_TOP_TOKENS_RESPONSE,
-    },
-
-    // 10. Providers V2 - Unified Buy (UAT only)
-    {
-      urlEndpoint:
-        /^https:\/\/on-ramp-cache\.uat-api\.cx\.metamask\.io\/v2\/regions\/[^/]+\/providers\?.*$/,
-      responseCode: 200,
-      response: RAMPS_PROVIDERS_RESPONSE,
-    },
-
-    // 11. Tokens (legacy path for backward compatibility) - UAT only
-    {
-      urlEndpoint:
-        /^https:\/\/on-ramp-cache\.uat-api\.cx\.metamask\.io\/regions\/[^/]+\/tokens\?.*$/,
-      responseCode: 200,
-      response: RAMPS_TOP_TOKENS_RESPONSE,
-    },
-
-    // 12. Payments V2 - per-provider payment methods (UAT only)
-    {
-      urlEndpoint:
-        /^https:\/\/on-ramp-cache\.uat-api\.cx\.metamask\.io\/v2\/regions\/[^/]+\/payments\?.*$/,
-      responseCode: 200,
-      response: RAMPS_PAYMENTS_V2_RESPONSE,
-    },
-
-    // FALLBACK: Catch-all for tokens endpoint with any region format (including malformed) - UAT only
-    // This handles legacy useRampTokens API calls where region might be incorrectly formatted
-    {
-      urlEndpoint:
-        /^https:\/\/on-ramp-cache\.uat-api\.cx\.metamask\.io\/regions\/.*\/tokens\?.*$/,
-      responseCode: 200,
-      response: RAMPS_TOP_TOKENS_RESPONSE,
-    },
-
-    {
-      urlEndpoint:
-        /^https:\/\/uat-static\.cx\.metamask\.io\/api\/v2\/tokenIcons\/assets\/.*\.png$/,
-      responseCode: 200,
-      response: '',
-    },
-  ];
-
-  await Promise.all([
-    ...mockEndpoints.map((mock) =>
-      setupMockRequest(mockServer, {
-        requestMethod: 'GET',
-        url: mock.urlEndpoint,
-        response: mock.response,
-        responseCode: mock.responseCode,
-      }),
-    ),
-    setupMockRequest(mockServer, {
-      requestMethod: 'HEAD',
-      url: /^https:\/\/uat-static\.cx\.metamask\.io\/api\/v2\/tokenIcons\/assets\/.*\.png$/,
-      response: '',
-      responseCode: 200,
-    }),
-  ]);
-};
