+++ tests/api-mocking/mock-responses/ramps/responses/ramps-networks-response.ts
@@ -1,0 +1,35 @@
+export const RAMPS_NETWORKS_RESPONSE = {
+  networks: [
+    {
+      active: true,
+      chainId: '1',
+      chainName: 'Ethereum Mainnet',
+      shortName: 'Ethereum',
+      nativeTokenSupported: true,
+    },
+    {
+      active: true,
+      chainId: '59144',
+      chainName: 'Linea',
+      shortName: 'Linea',
+      isEvm: true,
+      nativeTokenSupported: true,
+    },
+    {
+      active: true,
+      chainId: '8453',
+      chainName: 'Base',
+      shortName: 'Base',
+      isEvm: true,
+      nativeTokenSupported: true,
+    },
+    {
+      active: true,
+      chainId: 'solana:5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp',
+      chainName: 'Solana',
+      shortName: 'Solana',
+      nativeTokenSupported: true,
+      isEvm: false,
+    },
+  ],
+};
