+++ tests/api-mocking/mock-responses/ramps/responses/ramps-tokens-response.ts
@@ -1,0 +1,147 @@
+// Top tokens V2 response — available tokens across networks.
+export const RAMPS_TOP_TOKENS_RESPONSE = {
+  topTokens: [
+    {
+      assetId: 'eip155:1/erc20:0xaca92e438df0b2401ff60da7e4337b687a2435da',
+      chainId: 'eip155:1',
+      decimals: 6,
+      name: 'MetaMask USD',
+      symbol: 'MUSD',
+      iconUrl:
+        'https://static.cx.metamask.io/api/v2/tokenIcons/assets/eip155/1/erc20/0xaca92e438df0b2401ff60da7e4337b687a2435da.png',
+      tokenSupported: true,
+    },
+    {
+      assetId: 'eip155:59144/erc20:0xaca92e438df0b2401ff60da7e4337b687a2435da',
+      chainId: 'eip155:59144',
+      decimals: 6,
+      name: 'MetaMask USD',
+      symbol: 'MUSD',
+      iconUrl:
+        'https://static.cx.metamask.io/api/v2/tokenIcons/assets/eip155/59144/erc20/0xaca92e438df0b2401ff60da7e4337b687a2435da.png',
+      tokenSupported: true,
+    },
+    {
+      assetId: 'eip155:1/slip44:60',
+      chainId: 'eip155:1',
+      decimals: 18,
+      name: 'Ethereum',
+      symbol: 'ETH',
+      iconUrl:
+        'https://static.cx.metamask.io/api/v2/tokenIcons/assets/eip155/1/slip44/60.png',
+      tokenSupported: true,
+    },
+    {
+      assetId: 'bip122:000000000019d6689c085ae165831e93/slip44:0',
+      chainId: 'bip122:000000000019d6689c085ae165831e93',
+      decimals: 8,
+      name: 'Bitcoin',
+      symbol: 'BTC',
+      iconUrl:
+        'https://static.cx.metamask.io/api/v2/tokenIcons/assets/bip122/000000000019d6689c085ae165831e93/slip44/0.png',
+      tokenSupported: true,
+    },
+    {
+      assetId: 'solana:5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp/slip44:501',
+      chainId: 'solana:5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp',
+      decimals: 9,
+      name: 'Solana',
+      symbol: 'SOL',
+      iconUrl:
+        'https://static.cx.metamask.io/api/v2/tokenIcons/assets/solana/5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp/slip44/501.png',
+      tokenSupported: true,
+    },
+    {
+      assetId: 'eip155:56/slip44:714',
+      chainId: 'eip155:56',
+      decimals: 18,
+      name: 'Binance Coin',
+      symbol: 'BNB',
+      iconUrl:
+        'https://static.cx.metamask.io/api/v2/tokenIcons/assets/eip155/56/slip44/714.png',
+      tokenSupported: true,
+    },
+    {
+      assetId: 'eip155:8453/erc20:0x833589fcd6edb6e08f4c7c32d4f71b54bda02913',
+      chainId: 'eip155:8453',
+      decimals: 6,
+      name: 'USD Coin',
+      symbol: 'USDC',
+      iconUrl:
+        'https://static.cx.metamask.io/api/v2/tokenIcons/assets/eip155/8453/erc20/0x833589fcd6edb6e08f4c7c32d4f71b54bda02913.png',
+      tokenSupported: true,
+    },
+  ],
+  allTokens: [
+    {
+      assetId: 'eip155:1/erc20:0xaca92e438df0b2401ff60da7e4337b687a2435da',
+      chainId: 'eip155:1',
+      decimals: 6,
+      name: 'MetaMask USD',
+      symbol: 'MUSD',
+      iconUrl:
+        'https://static.cx.metamask.io/api/v2/tokenIcons/assets/eip155/1/erc20/0xaca92e438df0b2401ff60da7e4337b687a2435da.png',
+      tokenSupported: true,
+    },
+    {
+      assetId: 'eip155:59144/erc20:0xaca92e438df0b2401ff60da7e4337b687a2435da',
+      chainId: 'eip155:59144',
+      decimals: 6,
+      name: 'MetaMask USD',
+      symbol: 'MUSD',
+      iconUrl:
+        'https://static.cx.metamask.io/api/v2/tokenIcons/assets/eip155/59144/erc20/0xaca92e438df0b2401ff60da7e4337b687a2435da.png',
+      tokenSupported: true,
+    },
+    {
+      assetId: 'eip155:1/slip44:60',
+      chainId: 'eip155:1',
+      decimals: 18,
+      name: 'Ethereum',
+      symbol: 'ETH',
+      iconUrl:
+        'https://static.cx.metamask.io/api/v2/tokenIcons/assets/eip155/1/slip44/60.png',
+      tokenSupported: true,
+    },
+    {
+      assetId: 'bip122:000000000019d6689c085ae165831e93/slip44:0',
+      chainId: 'bip122:000000000019d6689c085ae165831e93',
+      decimals: 8,
+      name: 'Bitcoin',
+      symbol: 'BTC',
+      iconUrl:
+        'https://static.cx.metamask.io/api/v2/tokenIcons/assets/bip122/000000000019d6689c085ae165831e93/slip44/0.png',
+      tokenSupported: true,
+    },
+    {
+      assetId: 'solana:5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp/slip44:501',
+      chainId: 'solana:5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp',
+      decimals: 9,
+      name: 'Solana',
+      symbol: 'SOL',
+      iconUrl:
+        'https://static.cx.metamask.io/api/v2/tokenIcons/assets/solana/5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp/slip44/501.png',
+      tokenSupported: true,
+    },
+    {
+      assetId: 'eip155:56/slip44:714',
+      chainId: 'eip155:56',
+      decimals: 18,
+      name: 'Binance Coin',
+      symbol: 'BNB',
+      iconUrl:
+        'https://static.cx.metamask.io/api/v2/tokenIcons/assets/eip155/56/slip44/714.png',
+      tokenSupported: true,
+    },
+    {
+      assetId: 'eip155:8453/erc20:0x833589fcd6edb6e08f4c7c32d4f71b54bda02913',
+      chainId: 'eip155:8453',
+      decimals: 6,
+      name: 'USD Coin',
+      symbol: 'USDC',
+      iconUrl:
+        'https://static.cx.metamask.io/api/v2/tokenIcons/assets/eip155/8453/erc20/0x833589fcd6edb6e08f4c7c32d4f71b54bda02913.png',
+      tokenSupported: true,
+    },
+  ],
+};
